backend/requirements.txt
asgiref==3.6.0
Django==3.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
djangorestframework-simplejwt
black==22.6.0
flake8==4.0.1
coverage==6.5.0
coveralls==3.3.1
psycopg2-binary
django-ipware
django-storages
boto3
channels
daphne
channels_redis
pusher